Break up S3StorageClient
